用户批量删除提交

+++ cweb/src/components/page/UserInfoManage.vue
@@ -159,16 +159,35 @@
 					this.$message.error('请先选择用户');
 					return;
 				}
-				let ids = [];
-				this.selectedData.forEach((item) => {
-					ids.push(item.id);
-				})
-				this.$axios.delete("/v1/users/batch?ids=" + ids).then((response) => {
-					loading.close();
-				}, (response) => {
-					loading.close();
-					this.$message.error('批量删除用户失败');
-				});
+        this.$confirm('确认删除选中用户吗', '提示', {
+        	confirmButtonText: '确定',
+        	cancelButtonText: '取消',
+        	type: 'warning'
+        }).then(() => {
+        	let ids = [];
+        	this.selectedData.forEach((item) => {
+        		ids.push(item.id);
+        	})
+          const loading = this.$loading({
+          	lock: true,
+          	text: '加载中...',
+          	spinner: 'el-icon-loading',
+          	background: 'rgba(0, 0, 0, 0.7)'
+          });
+        	this.$axios.delete("/v1/users/batch?ids=" + ids).then((response) => {
+        		loading.close();
+            this.$message.success('批量删除用户成功');
+            this.findUsers();
+        	}, (response) => {
+        		loading.close();
+        		this.$message.error('批量删除用户失败');
+        	});
+        }).catch(() => {
+        	this.$message({
+        		type: 'info',
+        		message: '已取消删除'
+        	});
+        });
 			},
 			handleSizeChange(val) {
 				this.page.page_size = val;
